captcha_bypass/requirements.txt
keras
imutils
opencv-python
numpy
pytesseract
pillow
scikit-learn
